Move TestPlanner to internals
