+++ js/packages/teams-ai/src/planners/index.ts
@@ -10,4 +10,3 @@ export * from './ActionPlanner';
 export * as preview from './AssistantsPlanner';
 export * from './LLMClient';
 export * from './Planner';
-export * from './TestPlanner';
